Set the current status on dlrg.net
Expands dlrgclient.py to allow setting a Status class directly to
dlrg.net. This is used by Main to set the status when pressing the
state button. There will be no refresh if the state comes from dlrg
.net.

Makes the refresh_loop in Main a real loop so that the status will be
 refreshed regularly.

Adds multiple print-outs for debugging. dlrgclient.Status includes a
__str__ method to be printed beautifully.

Side note: This should have been at least three different commits.