Fix a bug where the remote update would run uncontrollably
Fix a bug in Main where it would on each refresh from remote try to
update the remote side even if the change has no effect. This happened
when the station was set to be not on duty. Then the app would detect
a significant change in temperature and would try to update with no
effect.